server log wip

+++ src/command/create.py
@@ -10,4 +10,7 @@ class CreateCommand(Command):
         if not os.path.exists(config_path):
             os.makedirs(config_path)
 
-        copyfile("conf/default.yaml", config_path + "/" + server.name + ".yaml")
+        config_file = config_path + "/" + server.name + ".yaml"
+        if not os.path.exists(config_file):
+            server.log.append("Creating server config " + server.name + " at " + config_file)
+            copyfile("conf/default.yaml", config_path + "/" + server.name + ".yaml")
